requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
httpx==0.27.0
python-dotenv==1.0.0
feedparser==6.0.10
meilisearch==0.20.0
websockets==12.0
requests==2.31.0
beautifulsoup4
pytest-mock
pytest-asyncio
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
aiofiles==23.2.1
locust==2.37.10
pytest-cov==6.2.1
